reorg and fix compiler error
